Refactor tests and update testing framework
Refactored the code by removing redundant ParagraphPortion test in ShapeCrawler.Tests.Unit.xUnit and moved it to ShapeCrawler.Tests.Unit. Also, the 'Fact' attribute in LineShape test has been updated to 'Test' reflecting a shift to NUnit from xUnit. This restructured and modernized testing structure enhances code readability and maintenance.
