+++ test/ShapeCrawler.Tests.Unit.xUnit/ParagraphPortionTests.cs
@@ -17,23 +17,6 @@ namespace ShapeCrawler.Tests.Unit;
 public class ParagraphPortionTests : SCTest
 {
     [Fact]
-    public void Text_Getter_returns_text_of_paragraph_portion()
-    {
-        // Arrange
-        var pptx = StreamOf("009_table");
-        var pres = new Presentation(pptx);
-        IParagraphPortion portion = ((ITable)pres.Slides[2].Shapes.First(sp => sp.Id == 3)).Rows[0].Cells[0]
-            .TextFrame
-            .Paragraphs[0].Portions[0];
-
-        // Act
-        string paragraphPortionText = portion.Text;
-
-        // Assert
-        paragraphPortionText.Should().BeEquivalentTo("0:0_p1_lvl1");
-    }
-
-    [Fact]
     public void Text_Setter_updates_text()
     {
         // Arrange
